requirements.txt
pandas
matplotlib
ta-lib
numpy
scikit-learn
tensorflow-cpu==2.8.0